requirements.txt
streamlit==1.38.0
pandas==2.2.3
google-generativeai==0.7.2
kaggle==1.6.17
